#map loading

+++ ExampleApp/ExampleApp/helpers/Constants.cs
@@ -4,6 +4,6 @@
     {
         public static int WidthElements = 13;
         public static int HeightElements = 7;
-        public static float SizeOfPlayer = 90.0f;
+        public static int SizeOfPlayer = 90;
     }
 }
